requirements.txt
# Nova Player — Python dependencies

# UI + Video
PyQt6>=6.6.0
python-vlc>=3.0.18122

# AI / ASR
faster-whisper>=1.0.0
soundfile>=0.12.1
numpy>=1.26.0
psutil>=5.9.0
sortedcontainers>=2.4.0

# Audio
ffmpeg-python>=0.2.0

# Optional — VAD (Silero)
# torch>=2.2.0
# torchaudio>=2.2.0

# Optional — SED (YAMNet)
# tensorflow>=2.15.0
# tensorflow-hub>=0.15.0

# Packaging
pyinstaller>=6.4.0
